+++ flow/declarations/process.js
@@ -1,10 +1,10 @@
 /* @flow */
 declare var process: { // eslint-disable-line
   env: {
-    AMQP_URL: string,
+    AMQP_URL?: string,
     APP_NAME: string,
     DATABASE_URL: string,
-    REDIS_URL: string,
+    REDIS_URL?: string,
     DISABLE_JWT_AUTH: string,
     DYNO: string,
     GIT_SHA: string,
@@ -14,8 +14,7 @@ declare var process: { // eslint-disable-line
     NODE_ENV: string,
     PORT: string,
     ROLLBAR_TOKEN: string,
-    SENDGRID_API_KEY: string,
-    SENDGRID_DEV_EMAIL: string,
+    SENDGRID_API_KEY?: string,
     SLACK_WEBHOOK_URL: string,
     CRYPTO_KEY?: string,
     CRYPTO_ALGO?: string,