+++ src/lib/constants.js
@@ -1,0 +1,5 @@
+// @flow
+
+export const SECOND = 1000;
+export const MINUTE = SECOND * 60;
+export const HOUR = MINUTE * 60;